+++ trl/trainer/__init__.py
@@ -1,3 +1,3 @@
 from .base import BaseTrainer
 from .utils import AdaptiveKLController, FixedKLController, LengthSampler
-from .accelerate_ppo import PPOTrainer
+from .ppo_trainer import PPOTrainer
